wip(#369): split status-bar/bottom-pin/protocol + shared test-helpers
