+++ src/cli/terminal-compositor.test.ts
@@ -5259,392 +5259,3 @@ describe('TerminalCompositor — picker mode', () => {
   });
 });
 
-describe('TerminalCompositor — background-status-bar coexistence', () => {
-  let stdout: MockStdout;
-  let stdin: MockStdin;
-  let writes: ReturnType<typeof collectWrites>;
-
-  beforeEach(() => {
-    stdout = makeMockStdout();
-    stdin = makeMockStdin();
-    writes = collectWrites(stdout);
-  });
-
-  /**
-   * Extract all CUP row numbers from an ANSI escape sequence string.
-   * CUP format: ESC [ <row> ; <col> H
-   */
-  function extractCupRows(out: string): number[] {
-    const rows: number[] = [];
-    const re = /\x1b\[(\d+);\d+H/g;
-    let m: RegExpExecArray | null;
-    while ((m = re.exec(out)) !== null) {
-      rows.push(parseInt(m[1]!, 10));
-    }
-    return rows;
-  }
-
-  it('keeps compositor frame above bg bar rows when extraRows > 0', async () => {
-    // rows=24, extraRows=2 → targetBottomRow = 24-1-2 = 21
-    // All CUP rows emitted by repaint() must be ≤ 21 so the compositor
-    // never writes into the two rows owned by BackgroundStatusBar.
-    const mockScrollRegion = {
-      withFullScrollRegion<T>(fn: () => T): T { return fn(); },
-      getExtraRows(): number { return 2; },
-    };
-    const c = new TerminalCompositor({
-      stdout,
-      stdin,
-      onCancel: vi.fn(),
-      scrollRegion: mockScrollRegion,
-    });
-    await c.arm();
-    writes.clear();
-
-    // Trigger a repaint by setting overlay content.
-    c.setOverlay('test overlay line');
-    const out = writes.all();
-
-    const cupRows = extractCupRows(out);
-    expect(cupRows.length).toBeGreaterThan(0);
-    const maxCupRow = Math.max(...cupRows);
-    // Must stay at or below targetBottomRow = 21 (rows=24, extraRows=2).
-    expect(maxCupRow).toBeLessThanOrEqual(21);
-    c.disarm();
-  });
-
-  it('uses full bottom row when extraRows is 0', async () => {
-    // rows=24, extraRows=0 → targetBottomRow = 24-1-0 = 23
-    // The compositor should use the full available space (no reserved rows).
-    const mockScrollRegion = {
-      withFullScrollRegion<T>(fn: () => T): T { return fn(); },
-      getExtraRows(): number { return 0; },
-    };
-    const c = new TerminalCompositor({
-      stdout,
-      stdin,
-      onCancel: vi.fn(),
-      scrollRegion: mockScrollRegion,
-    });
-    await c.arm();
-    writes.clear();
-
-    c.setOverlay('test overlay line');
-    const out = writes.all();
-
-    const cupRows = extractCupRows(out);
-    expect(cupRows.length).toBeGreaterThan(0);
-    const maxCupRow = Math.max(...cupRows);
-    // Should reach the full bottom row = 23 (rows=24, extraRows=0).
-    expect(maxCupRow).toBeLessThanOrEqual(23);
-    // Also verify it actually uses the bottom (not unnecessarily clamped).
-    expect(maxCupRow).toBeGreaterThanOrEqual(22);
-    c.disarm();
-  });
-});
-
-describe('TerminalCompositor — input bottom-pin placement', () => {
-  // These tests verify that the live input frame is ALWAYS bottom-pinned
-  // (targetBottomRow === absoluteBottom) — on a fresh session, with or without
-  // a banner, and after any number of commits.
-  //
-  // Core invariant: the input line is the last frameLines entry and always
-  // lands on absoluteBottom (rows-1-extraRows); the dropdown / hint / streaming
-  // overlay grow UPWARD into the empty viewport above it. This is what lets the
-  // slash-command completion menu open on a brand-new session without shoving
-  // the prompt down to make headroom.
-  //
-  // History: this used to be a two-regime "content-following" placement — the
-  // frame pinned just below the banner at
-  //   targetBottomRow = min(absoluteBottom, max(anchorRow, committedBandBottomRow) + physicalRows)
-  // while idle with a banner, marching down to absoluteBottom only as committed
-  // content accumulated. That left a fresh-session prompt one row under the
-  // banner with no headroom, so opening the dropdown grew physicalRows and
-  // pushed the whole frame down. The regime was removed in favour of
-  // unconditional bottom-pinning; the banner is still protected as a ceiling by
-  // the anchorRow floor (frame-preserve.ts / committed-band-repin.ts).
-
-  let stdout: MockStdout;
-  let stdin: MockStdin;
-  let writes: ReturnType<typeof collectWrites>;
-
-  beforeEach(() => {
-    stdout = makeMockStdout();
-    stdin = makeMockStdin();
-    writes = collectWrites(stdout);
-  });
-
-  it('cold start: no banner, no content — frame stays bottom-pinned', async () => {
-    // Before any commit and without a banner, the frame must land at the
-    // standard bottom row (rows-1).
-    stdout.rows = 70;
-    const c = new TerminalCompositor({ stdout, stdin, onCancel: vi.fn(), promptText: '> ' });
-    await c.arm();
-    const out = writes.all();
-    // 1-line idle frame → bottom-pinned at row 69 (70-1).
-    expect(out).toContain('\x1b[69;1H');
-    c.disarm();
-  });
-
-  it('tall terminal + banner: frame stays bottom-pinned (no content-following) so the dropdown has headroom', async () => {
-    // Regression guard for the fresh-session dropdown-jump fix: rows=70,
-    // anchorRow=15 (14-row welcome banner). Even with a small committed band
-    // sitting near the banner (committedBandBottomRow=16, far above
-    // absoluteBottom=69), the next standalone repaint must bottom-pin the input
-    // frame — NOT follow the content up to ~row 19. Bottom-pinning is what
-    // leaves the empty viewport above the prompt for the completion dropdown to
-    // grow into without shoving the input down.
-    stdout.rows = 70;
-    const c2 = new TerminalCompositor({
-      stdout, stdin, onCancel: vi.fn(), promptText: '> ',
-      anchorRow: 15,
-    });
-    await c2.arm();
-    // Manually force a small committedBandBottomRow by patching internal state
-    // via the typed cast the compositor tests already use.
-    const internals = c2 as unknown as {
-      committedBand: string[];
-      committedBandTopRow: number;
-      committedBandBottomRow: number;
-      hasCommitted: boolean;
-      logUpdate: { resetGeometry?: () => void };
-    };
-    internals.committedBand = ['COMMITTED'];
-    internals.committedBandTopRow = 16;
-    internals.committedBandBottomRow = 16;
-    internals.hasCommitted = true;
-    // Reset CupFrameRenderer geometry so its erase pass on the next render
-    // doesn't re-visit the stale previous-frame row (row 69 from arm()).
-    internals.logUpdate.resetGeometry?.();
-    writes.clear();
-    c2.setOverlay('FOLLOW_TEST');
-    const out2 = writes.all();
-    // Input frame must land at absoluteBottom = row 69 (70-1), NOT follow the
-    // band up to ~row 19.
-    expect(out2).toContain('\x1b[69;1H');
-    // The overlay text must appear in the output (frame rendered).
-    expect(out2).toContain('FOLLOW_TEST');
-    c2.disarm();
-  });
-
-  it('no-banner session: frame stays bottom-pinned regardless of committed content', async () => {
-    // Without a banner (anchorRow undefined or ≤1) the frame is always at
-    // absoluteBottom = rows-1-extraRows regardless of how many commits have
-    // accumulated — this preserves all resize-ghost, shrink-gap, and
-    // scrollback-gap invariants.
-    stdout.rows = 24;
-    const c = new TerminalCompositor({ stdout, stdin, onCancel: vi.fn(), promptText: '> ' });
-    await c.arm();
-
-    // Commit several lines — with no banner the frame should stay bottom-pinned.
-    for (let i = 0; i < 5; i++) {
-      c.commitAbove(`LINE_${i}`);
-    }
-    writes.clear();
-    c.setOverlay('AFTER_COMMITS');
-    const out = writes.all();
-
-    // Frame must always be at absoluteBottom = rows-1 = 23, regardless of
-    // committed content.
-    expect(out).toContain('\x1b[23;1H');
-    c.disarm();
-  });
-
-  it('reserved extraRows: targetBottomRow never enters reserved rows', async () => {
-    // extraRows=2 → absoluteBottom = 24-1-2 = 21.
-    // With a banner (anchorRow=5) and committed content, the bottom-pinned frame
-    // must cap at absoluteBottom=21 — never write into bg-status-bar rows 22-23.
-    const mockScrollRegion = {
-      withFullScrollRegion<T>(fn: () => T): T { return fn(); },
-      getExtraRows(): number { return 2; },
-    };
-    stdout.rows = 24;
-    const c = new TerminalCompositor({
-      stdout, stdin, onCancel: vi.fn(), promptText: '> ',
-      anchorRow: 5,
-      scrollRegion: mockScrollRegion,
-    });
-    await c.arm();
-    // Commit enough lines to fill the viewport; the frame stays bottom-pinned.
-    for (let i = 0; i < 25; i++) {
-      c.commitAbove(`LINE_${i}`);
-    }
-    writes.clear();
-    c.setOverlay('EXTRA_ROW_TEST');
-    const out = writes.all();
-    // Collect CUP rows, excluding the eviction-scroll CUP at physicalBottom=24.
-    // evictRowsToScrollback writes `\x1b[24;1H\n...` to trigger DECSTBM scroll;
-    // that row is intentionally at the physical margin (not a frame content row).
-    const physicalBottom = stdout.rows; // 24
-    const re = /\x1b\[(\d+);\d+H/g;
-    let m: RegExpExecArray | null;
-    let maxFrameRow = 0;
-    while ((m = re.exec(out)) !== null) {
-      const row = parseInt(m[1]!, 10);
-      if (row !== physicalBottom) {
-        maxFrameRow = Math.max(maxFrameRow, row);
-      }
-    }
-    expect(maxFrameRow).toBeGreaterThan(0);
-    // Frame content must stay at or below absoluteBottom = 21 (extraRows=2).
-    expect(maxFrameRow).toBeLessThanOrEqual(21);
-    c.disarm();
-  });
-});
-
-// ─── Protocol invariants ─────────────────────────────────────────────────────
-//
-// These tests guard the externally-governed contracts catalogued in
-// docs/tui-invariants.md. Each test names the historical bug it prevents from
-// recurring. If you find yourself disabling one of these, you are also opting
-// out of the invariant — confirm in the PR description that the contract has
-// genuinely changed at the protocol level (VT spec, log-update source) before
-// merging.
-
-describe('TerminalCompositor — protocol invariants', () => {
-  let stdout: MockStdout;
-  let stdin: MockStdin;
-  let writes: ReturnType<typeof collectWrites>;
-
-  beforeEach(() => {
-    stdout = makeMockStdout();
-    stdin = makeMockStdin();
-    writes = collectWrites(stdout);
-  });
-
-  describe('arm() — log-update anchor protocol', () => {
-    it('writes CUP to bottom row before the first repaint', async () => {
-      // External constraint (log-update anchor protocol): the FIRST log-update
-      // render writes at the current cursor row with no preceding cursor
-      // movement; all subsequent repaints anchor relative to that row.
-      //
-      // Prevents recurrence of ce1dcfe: without the explicit CUP to
-      // (rows-1, 1), the overlay anchors wherever the welcome banner +
-      // status-line save/restore left the cursor (mid-screen) — producing
-      // a frame split where the overlay paints at the top and input
-      // drifts to the bottom.
-      const c = new TerminalCompositor({ stdout, stdin, onCancel: vi.fn() });
-      await c.arm();
-      const out = writes.all();
-      const anchorRow = (stdout.rows ?? 24) - 1;
-      const expectedCUP = `\x1b[${anchorRow};1H`;
-      const cupIdx = out.indexOf(expectedCUP);
-      expect(cupIdx).toBeGreaterThanOrEqual(0);
-      // CUP must precede any log-update frame content. The first frame
-      // includes the input-row caret render. We assert the CUP appears
-      // within the first 64 bytes of arm()'s output to keep the test
-      // robust against minor preamble changes (e.g. raw-mode setup
-      // sequences) without losing the "comes first" guarantee.
-      expect(cupIdx).toBeLessThan(64);
-    });
-
-    it('skips the CUP write when stdout is not a TTY', async () => {
-      // Negative complement: non-TTY surfaces (pipes, CI) have no cursor
-      // to position. Writing the CUP would inject literal escape bytes
-      // into piped output — visible garbage in logs.
-      const nonTTY = makeMockStdout(false);
-      const nonTTYWrites = collectWrites(nonTTY);
-      const c = new TerminalCompositor({
-        stdout: nonTTY,
-        stdin,
-        onCancel: vi.fn(),
-      });
-      await c.arm();
-      // No CUP escape should appear (the regex matches `\x1b[<n>;1H`).
-      expect(nonTTYWrites.all()).not.toMatch(/\x1b\[\d+;1H/);
-    });
-  });
-
-  describe('commitAbove() — DECSTBM contract', () => {
-    it('routes the newline write through withFullScrollRegion when scrollRegion is provided', async () => {
-      // External constraint (DECSTBM contract): when a status line is active,
-      // the bottom row is reserved via a persistent scroll region. A raw `\n`
-      // at the bottom of that sub-region triggers a sub-region scroll and the
-      // displaced top line silently exits without entering scrollback.
-      //
-      // Prevents recurrence of f962403-family bugs: removing the
-      // withFullScrollRegion guard would resurface "tool-lane lines vanish
-      // into the void" the moment the status line is active.
-      const scrollRegion = {
-        withFullScrollRegion: vi.fn(<T,>(fn: () => T): T => fn()),
-        // getExtraRows is part of the CompositorScrollRegionGuard contract:
-        // arm()→repaint() reads it to keep the frame above any reserved
-        // status-bar rows. A mock that omits it throws on arm(); 0 means
-        // "no reserved rows", isolating this test to the routing invariant.
-        getExtraRows: vi.fn(() => 0),
-      };
-      const c = new TerminalCompositor({
-        stdout,
-        stdin,
-        onCancel: vi.fn(),
-        scrollRegion,
-      });
-      await c.arm();
-      writes.clear();
-      c.commitAbove('COMMITTED_BLOCK');
-
-      // The guard-call is the load-bearing assertion: removing
-      // `writeWithGuard(...)` from commitAbove would make this fail
-      // immediately, regardless of how the inner write is shaped.
-      expect(scrollRegion.withFullScrollRegion).toHaveBeenCalled();
-      // Content reaches stdout. Armed-path commitAbove (CommittedBand.commitAbove)
-      // positions the cursor at the row above the frame and erases the line
-      // before writing the block, so the committed text appears as
-      // `CUP-to-row → EL (\x1b[2K) → text`. We pin that exact positioned
-      // shape rather than a bare substring so the test fails if the block
-      // ever stops being placed via the centralized cursor-positioned write.
-      const out = writes.all();
-      expect(out).toContain('COMMITTED_BLOCK');
-      expect(out).toMatch(/\x1b\[\d+;1H\x1b\[2KCOMMITTED_BLOCK/);
-    });
-
-    it('writes directly without invoking the guard when scrollRegion is absent', async () => {
-      // Negative complement: callers without an active status line shouldn't
-      // pay for indirection; the bare write still commits to scrollback
-      // because no DECSTBM is in effect.
-      const c = new TerminalCompositor({ stdout, stdin, onCancel: vi.fn() });
-      await c.arm();
-      writes.clear();
-      c.commitAbove('NO_GUARD_BLOCK');
-      // Same positioned emission shape as the guarded path — armed
-      // commitAbove writes `CUP-to-row → EL (\x1b[2K) → text`. The only
-      // difference from the guarded case is that withFullScrollRegion is
-      // never invoked (asserted implicitly: no scrollRegion was provided).
-      const out = writes.all();
-      expect(out).toContain('NO_GUARD_BLOCK');
-      expect(out).toMatch(/\x1b\[\d+;1H\x1b\[2KNO_GUARD_BLOCK/);
-    });
-  });
-
-  describe('disarm() — cursor visibility restore', () => {
-    it('calls logUpdate.done() after logUpdate.clear() so the cursor stays visible', async () => {
-      // External constraint (cliCursor visibility): log-update hides the
-      // cursor on every render() when showCursor is false (the default).
-      // Only done() calls cliCursor.show(); clear() alone leaves the cursor
-      // hidden, leaking that state for the rest of the session.
-      //
-      // Prevents recurrence of "cursor stays invisible after a turn ends"
-      // by asserting the (clear, done) sequence is intact.
-      const c = new TerminalCompositor({ stdout, stdin, onCancel: vi.fn() });
-      await c.arm();
-
-      const internals = c as unknown as {
-        logUpdate: { clear: () => void; done: () => void };
-      };
-      const calls: string[] = [];
-      const originalClear = internals.logUpdate.clear.bind(internals.logUpdate);
-      const originalDone = internals.logUpdate.done.bind(internals.logUpdate);
-      internals.logUpdate.clear = () => {
-        calls.push('clear');
-        originalClear();
-      };
-      internals.logUpdate.done = () => {
-        calls.push('done');
-        originalDone();
-      };
-
-      c.disarm();
-      expect(calls).toEqual(['clear', 'done']);
-    });
-  });
-});
